Fix squid:S1214: Constants should not be defined in interfaces Move TransactionManager.TRANSACTION_ID_RESOURCE_NAME to TransactionConstants.TRANSACTION_ID_RESOURCE_NAME
