+++ molgenis-data/src/main/java/org/molgenis/data/transaction/TransactionConstants.java
@@ -1,0 +1,7 @@
+package org.molgenis.data.transaction;
+
+public final class TransactionConstants {
+  public static final String TRANSACTION_ID_RESOURCE_NAME = "transactionId";
+
+  private TransactionConstants(){}
+}
