+++ molgenis-data/src/main/java/org/molgenis/data/transaction/TransactionManager.java
@@ -3,7 +3,6 @@ package org.molgenis.data.transaction;
 import org.springframework.transaction.PlatformTransactionManager;
 
 public interface TransactionManager extends PlatformTransactionManager {
-  String TRANSACTION_ID_RESOURCE_NAME = "transactionId";
 
   void addTransactionListener(TransactionListener transactionListener);
 }
